hr-chatbot-python/requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
pyodbc>=5.0.0

# AI/ML
sentence-transformers>=2.2.0
chromadb>=0.4.0
scikit-learn>=1.3.0
numpy>=1.26.0

# LLM for AI SQL Generator
openai>=1.0.0
groq>=0.4.0

# Utilities
python-multipart>=0.0.6
